+++ extension/chrome/settings/modules/contacts.ts
@@ -8,7 +8,7 @@ import { Att } from '../../../js/common/core/att.js';
 import { Browser } from '../../../js/common/browser/browser.js';
 import { Ui } from '../../../js/common/browser/ui.js';
 import { BrowserMsg } from '../../../js/common/browser/browser-msg.js';
-import { Pgp } from '../../../js/common/core/pgp-msg.js';
+import { Pgp } from '../../../js/common/core/pgp.js';
 import { Buf } from '../../../js/common/core/buf.js';
 import { AttUI } from '../../../js/common/ui/att_ui.js';
 import { KeyImportUi } from '../../../js/common/ui/key_import_ui.js';
